+++ packages/toolbars-v3/web/src/icons/ReplaceIcon.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const ReplaceIcon = (props: any) => (
+  <svg width="14" height="16" viewBox="0 0 14 16" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      // eslint-disable-next-line max-len
+      d="M1.91888 3.35901L3.56788 5.01002L2.85988 5.71602L0.00488281 2.85601L2.89888 0L3.60188 0.711002L1.93288 2.35901H7.00388C10.8609 2.35901 13.9999 5.49802 13.9999 9.35503H12.9999C12.9999 6.04802 10.3109 3.35901 7.00388 3.35901H1.91888ZM12.1179 12.3594L10.4339 10.6754L11.1399 9.96738L13.9939 12.8214L11.1399 15.6754L10.4329 14.9684L12.0419 13.3594H6.99594C3.13897 13.3594 0 10.2184 0 6.35938H0.999991C0.999991 9.66838 3.68897 12.3594 6.99594 12.3594H12.1179Z"
+      fill="black"
+    />
+  </svg>
+);
+
+export default ReplaceIcon;