+++ packages/toolbars-v3/web/src/pluginToolbarItemRenders.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ToggleButton, SettingsButton, LinkButton } from './components/buttons';
+import { ToggleButton, SettingsButton, LinkButton, ReplaceButton } from './components/buttons';
 import { PLUGIN_TOOLBAR_BUTTON_ID } from 'wix-rich-content-editor-common';
 
 export const pluginToolbarItemsRenders = {
@@ -23,6 +23,11 @@ export const pluginToolbarItemsRenders = {
   },
   [PLUGIN_TOOLBAR_BUTTON_ID.SETTINGS]: toolbarItem => {
     return <SettingsButton toolbarItem={toolbarItem} />;
+  },
+  [PLUGIN_TOOLBAR_BUTTON_ID.REPLACE]: toolbarItem => {
+    return (
+      <ReplaceButton toolbarItem={toolbarItem} onClick={args => toolbarItem.commands.click(args)} />
+    );
   },
   [PLUGIN_TOOLBAR_BUTTON_ID.LINK]: toolbarItem => {
     return <LinkButton toolbarItem={toolbarItem} />;